% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_number_missing.R
\name{count_number_missing}
\alias{count_number_missing}
\title{Count number of missing values}
\usage{
count_number_missing(x)
}
\arguments{
\item{x}{Vector, list, or data frame}
}
\value{
a vector containing the number of missing values for each column
}
\description{
\code{count_num_missing} counts the number of NAs in a vector, list,
or data frame (column-wise)
}
\examples{
x <- data.frame(a = c(1, NA, 3), b = c(4, NA, NA), c = c(7, NA, 9))
count_num_missing(x)
}
