% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_number_range.R
\name{verify_number_range}
\alias{verify_number_range}
\title{Verify that values are within a range}
\usage{
verify_number_range(x, lower, upper)
}
\arguments{
\item{x}{Numeric vector.}

\item{lower}{Number. The bottom threshold}

\item{upper}{Number. The upper threshold}
}
\value{
Boolean. Indicates if all values in vectors are within the range.
}
\description{
\code{verify_number_rangge} checks that all the values in a vector are within
the specified range
}
\examples{
verify_number_range(c(1, 12, 10000), lower = 1, upper = 100)
verify_number_range(c(-50, 12, 10000), lower = 1, upper = 100)
}
